+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.1] - 2025-08-08
+
+- Correct the blocks for `blockquote`.
+- Correct the blocks for `table of content`.
+
+
 ## [3.0.0] - 2025-08-08
 
 - Add a scope for included `metadata`.